# [opt](function) Optimize the concat(col, constant, constant, constant) function