--- a/be/src/vec/functions/function_string.h
+++ b/be/src/vec/functions/function_string.h
@@ -1038,9 +1038,15 @@ public:
             const auto* column_string = context->get_constant_col(i);
             if (column_string == nullptr) {
                 state->use_state = false;
-                break;
+                return IFunction::open(context, scope);
             }
             auto string_vale = column_string->column_ptr->get_data_at(0);
+            if (string_vale.data == nullptr) {
+                // For concat(col, null), it is handled by default_implementation_for_nulls
+                state->use_state = false;
+                return IFunction::open(context, scope);
+            }
+
             state->tail.append(string_vale.begin(), string_vale.size);
         }