## Proposed changes

```
mysql [test]>select count(concat(short, "123","121231","123123",'12312313')) from strings;
+-------------------------------------------------------------+
| count(concat(short, '123', '121231', '123123', '12312313')) |
+-------------------------------------------------------------+
|                                                    10000000 |
+-------------------------------------------------------------+
1 row in set (0.52 sec)

mysql [test]>select count(concat(short, "123","121231","123123",'12312313' , short , short, short)) from strings;
+----------------------------------------------------------------------------------+
| count(concat(short, '123', '121231', '123123', '12312313', short, short, short)) |
+----------------------------------------------------------------------------------+
|                                                                         10000000 |
+----------------------------------------------------------------------------------+
1 row in set (0.98 sec)



now

mysql [test]>select count(concat(short, "123","121231","123123",'12312313')) from strings;
+-------------------------------------------------------------+
| count(concat(short, '123', '121231', '123123', '12312313')) |
+-------------------------------------------------------------+
|                                                    10000000 |
+-------------------------------------------------------------+
1 row in set (0.19 sec)

mysql [test]>select count(concat(short, "123","121231","123123",'12312313' , short , short, short)) from strings;
+----------------------------------------------------------------------------------+
| count(concat(short, '123', '121231', '123123', '12312313', short, short, short)) |
+----------------------------------------------------------------------------------+
|                                                                         10000000 |
+----------------------------------------------------------------------------------+
1 row in set (0.71 sec)
```

<!--Describe your changes.-->